+++ packages/agtree/test/converter/network/network-rule-converter.test.ts
@@ -304,6 +304,20 @@ describe('NetworkRuleConverter', () => {
                     '||cdn.cookielaw.org^$important,redirect=noopjs,script,domain=open.spotify.com',
                 ],
             },
+
+            // should convert googlesyndication.com/adsbygoogle.js to googlesyndication-adsbygoogle
+            {
+                actual: '||example.com/adsbygoogle.js^$script,redirect-rule=googlesyndication.com/adsbygoogle.js',
+                expected: [
+                    '||example.com/adsbygoogle.js^$script,redirect-rule=googlesyndication-adsbygoogle',
+                ],
+            },
+            {
+                actual: '||example.com/adsbygoogle.js$script,redirect=googlesyndication.com/adsbygoogle.js,important',
+                expected: [
+                    '||example.com/adsbygoogle.js$script,redirect=googlesyndication-adsbygoogle,important',
+                ],
+            },
         ])('should convert \'$actual\' to \'$expected\'', (testData) => {
             expect(testData).toBeConvertedProperly(NetworkRuleConverter, 'convertToAdg');
         });